// src/types/cc-stats/agbypilot-attributes.ts
export enum StatsAgentByPilotAttributes {
    /** Number of calls received for this pilot. */
    nbCallsReceived = 'nbCallsReceived',

    /** Number of calls received by transfer. */
    nbCallsTransfIn = 'nbCallsTransfIn',

    /** Number of calls served. */
    nbCallsServed = 'nbCallsServed',

    /** Number of calls served too quickly. */
    nbCallsServedTooQuickly = 'nbCallsServedTooQuickly',

    /** Number of calls with enquiry. */
    nbCallsWithEnquiry = 'nbCallsWithEnquiry',

    /** Number of calls where help was requested. */
    nbCallsWithHelp = 'nbCallsWithHelp',

    /** Number of calls transferred from the agent. */
    nbCallsTransf = 'nbCallsTransf',

    /** Number of calls transferred to the agent. */
    nbCallsTransfToAgent = 'nbCallsTransfToAgent',

    /** Number of calls in wrap-up. */
    nbCallsInWrapup = 'nbCallsInWrapup',

    /** Maximum duration of call processing (hh:mm:ss). */
    maxCallProcDur = 'maxCallProcDur',

    /** Maximum duration of conversation (hh:mm:ss). */
    maxConvDur = 'maxConvDur',

    /** Maximum duration of wrap-up (hh:mm:ss). */
    maxWrapupDur = 'maxWrapupDur',

    /** Total duration of call processing (hh:mm:ss). */
    callProcTDur = 'callProcTDur',

    /** Average duration of call processing (hh:mm:ss). */
    callProcADur = 'callProcADur',

    /** Total duration of conversation (hh:mm:ss). */
    convTDur = 'convTDur',

    /** Average duration of conversation (hh:mm:ss). */
    convADur = 'convADur',

    /** Total duration of wrap-up (hh:mm:ss). */
    wrapupTDur = 'wrapupTDur',

    /** Average duration of wrap-up (hh:mm:ss). */
    wrapupADur = 'wrapupADur',

    /** Total duration of conversation occurring during wrap-up (hh:mm:ss). */
    convInWrapupTDur = 'convInWrapupTDur',

    /** Total busy time during wrap-up (hh:mm:ss). */
    busyTimeInWrapupTDur = 'busyTimeInWrapupTDur',

    /** Total duration of calls on hold (hh:mm:ss). */
    onHoldTDur = 'onHoldTDur',

    /** Average duration of calls on hold (hh:mm:ss). */
    onHoldADur = 'onHoldADur',

    /** Total duration of transaction phase (hh:mm:ss). */
    transTDur = 'transTDur',

    /** Average duration of transaction phase (hh:mm:ss). */
    transADur = 'transADur',

    /** Total duration of pause (hh:mm:ss). */
    pauseTDur = 'pauseTDur',

    /** Average duration of pause (hh:mm:ss). */
    pauseADur = 'pauseADur',

    /** All attributes (for testing purposes only; not for production use). */
    ALL = 'ALL',
}